backend/requirements.txt
gunicorn==20.0.4
Flask==1.1.1
requests==2.23.0
google-cloud-pubsub==1.4.3
google-cloud-storage==1.27.0
flask-cors==3.0.8